Renamed to FileCatalog
